Move core components into their own folder
