+++ core-components/app/index.js
@@ -5,9 +5,9 @@ import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
 import { withStyles } from 'critical-style-loader/lib';
 import favicon from 'assets/images/favicon.ico';
-import ConnectedRoot from 'components/connectedRoot';
-import ErrorBoundary from 'components/errorBoundary';
-import ErrorMessage from 'components/errorMessage';
+import ConnectedRoot from 'core-components/connectedRoot';
+import ErrorBoundary from 'core-components/errorBoundary';
+import ErrorMessage from 'core-components/errorMessage';
 import getRoots from 'selectors/getRoots';
 import Menu from 'assets/icons/menu.svg';
 import styles from './app.css';
