+++ core-components/connectedRoot/index.js
