+++ core-components/errorBoundary/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import ErrorMessage from 'components/errorMessage';
+import ErrorMessage from 'core-components/errorMessage';
 
 class ErrorBoundary extends React.Component {
   static propTypes = {
