+++ core-components/form/inputText.js
