+++ core-components/head/index.js
