+++ core-components/image/img.js
