+++ core-components/image/index.js
