+++ core-components/image/picture.js
