+++ core-components/placeholder/placeholder.css
