+++ core-components/placeholderLoading/index.js
