+++ core-components/placeholderLoading/placeholderLoading.css
